// types/global.d.ts
export type Articles = {
  id: string;
  title: string;
  image: string;
  date: { seconds: number; nanoseconds: number };
  author: string;
  category: string;
};

export type Categories = {
  id: string;
  name: string;
};